Move Capabilities to `dd_constants.py` and restore `remoteconfig.py` file where it was
