+++ tests/parametric/test_dynamic_configuration.py
@@ -8,7 +8,7 @@ import pytest
 from ddapm_test_agent.trace import root_span
 
 from utils import bug, context, features, irrelevant, missing_feature, rfc, scenarios, flaky
-from utils.remoteconfig import Capabilities
+from utils.dd_constants import Capabilities
 from utils.parametric.spec.trace import (
     Span,
     assert_trace_has_tags,
