+++ utils/dd_constants.py
@@ -33,3 +33,28 @@ class RemoteConfigApplyState(IntEnum):
     UNACKNOWLEDGED = 1
     ACKNOWLEDGED = 2
     ERROR = 3
+
+
+class Capabilities(IntEnum):
+    ASM_ACTIVATION = 1
+    ASM_IP_BLOCKING = 2
+    ASM_DD_RULES = 3
+    ASM_EXCLUSIONS = 4
+    ASM_REQUEST_BLOCKING = 5
+    ASM_ASM_RESPONSE_BLOCKING = 6
+    ASM_USER_BLOCKING = 7
+    ASM_CUSTOM_RULES = 8
+    ASM_CUSTOM_BLOCKING_RESPONSE = 9
+    ASM_TRUSTED_IPS = 10
+    ASM_API_SECURITY_SAMPLE_RATE = 11
+    APM_TRACING_SAMPLE_RATE = 12
+    APM_TRACING_LOGS_INJECTION = 13
+    APM_TRACING_HTTP_HEADER_TAGS = 14
+    APM_TRACING_CUSTOM_TAGS = 15
+    APM_TRACING_ENABLED = 19
+    ASM_RASP_SQLI = 21
+    ASM_RASP_LFI = 22
+    ASM_RASP_SSRF = 23
+    ASM_RASP_SHI = 24
+    APM_TRACING_SAMPLE_RULES = 29
+
